Collections API as separate module.
