+++ flower-core/src/main/java/org/xdcrafts/flower/core/impl/extensions/PredicateExtension.java
@@ -23,7 +23,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.function.Predicate;
 
-import static org.xdcrafts.flower.core.utils.MapDsl.Mutable.with;
+import static org.xdcrafts.flower.tools.MapDsl.Mutable.with;
 
 /**
  * Predicate-based implementation of Extension.
