+++ flower-core/src/main/java/org/xdcrafts/flower/core/impl/switches/KeywordSwitch.java
@@ -25,7 +25,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static org.xdcrafts.flower.core.utils.MapApi.DotNotation.dotGetString;
+import static org.xdcrafts.flower.tools.MapApi.DotNotation.dotGetString;
 
 /**
  * Implementation of Switch that selects Action based on value of keyword in context.
