+++ flower-spring/src/test/java/org/xdcrafts/flower/core/spring/example/Receiver.java
@@ -19,7 +19,7 @@ package org.xdcrafts.flower.core.spring.example;
 import java.util.HashMap;
 import java.util.Map;
 
-import static org.xdcrafts.flower.core.utils.MapDsl.Mutable.with;
+import static org.xdcrafts.flower.tools.MapDsl.Mutable.with;
 
 /**
  * Receiver.
